tests: refactor test strcuture
Now as the application increase in size and complexity our tests will be more easy to extend too
